refactor: move, end 입력 책임을 view로 변경
